requirements.txt
Flask==2.3.2
Flask-Session==0.4.0
python-dotenv==1.0.0
PyJWT
